UI and Bug Fixes
Details: Fixed issue where a player would hit start and depoly to the game would bring other players to the game in the wrong State. Fixed the UI for the ammo to count down correctly. Fixed some scaling issues with the UI that will still need further testing. Updated the UI to have the color select option enabled.