Started fixing sprite bugs with abilities
fixed Britney, Jo, and Anna's sprites and skills to function properly, and started a rewrite on Mio's ability functionality so it's not just a blind teleport. *started as in it's not done yet. Doesn't break the game, but doesn't have any real functionality yet, and mio becomes useless after doing it